requirements.txt
Flask
Pillow
gunicorn
requests
# AI Libraries
diffusers
transformers
accelerate
torch
torchvision
torchaudio
